test output fixes

+++ test/process.cpp
@@ -1142,7 +1142,7 @@ TEST_F(sys_call_test, procinfo_two_processchilds_cpuload)
 
 					uint64_t delta = tcpu - lastcpu;
 
-					printf("%d:%d)%ld:%ld)%ld >> %ld\n", (int)callnum, (int)ctid, tinfo->m_pid, tinfo->m_tid, tcpu, delta);
+					printf("%d:%d)%ld:%ld)%ld >> %ld\n", (int)callnum, (int)ctid, (int)tinfo->m_pid, (int)tinfo->m_tid, (int)tcpu, (int)delta);
 
 					if(callnum > 2)
 					{
